handled delete request errors

+++ backend/OurFootprint/scripts/dels.py
@@ -1,16 +1,25 @@
-from calculator.models import FortisBillField, HydroBillField, Commute
+from calculator.models import FortisBillField, HydroBillField, Commute, UserCommute
 
 
-def del_fortis(fortis_id):
-    instance = FortisBillField.objects.get(id=fortis_id)
+def del_fortis(uid, fortis_id):
+    instance = FortisBillField.objects.get(id=fortis_id, user_id=uid)
+    if instance is None:
+        return {"error": "not found"}, 404
+    instance.delete()
+    return
+
+
+def del_hydro(uid, hydro_id):
+    instance = HydroBillField.objects.get(id=hydro_id, user_id=uid)
+    if instance is None:
+        return {"error": "not found"}, 404
     instance.delete()
 
 
-def del_hydro(hydro_id):
-    instance = HydroBillField.objects.get(id=hydro_id)
-    instance.delete()
+def del_commute(uid, commute_id):
+    user_commute_ref = UserCommute.objects.get(user_id=uid, commute_id=commute_id)
+    if user_commute_ref is None:
+        return {"error": "not found"}, 404
 
-
-def del_commute(commute_id):
     instance = Commute.objects.get(commute_id=commute_id)
     instance.delete()
